# server/openapi.yaml
openapi: 3.0.3
info:
  title: Secretly API
  description: |
    Secure secrets management API with RBAC, audit logging, and comprehensive secret lifecycle management.
    
    ## Authentication
    All API endpoints (except health check) require Bearer token authentication.
    Include the token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    API requests are rate limited. Check response headers for current limits.
    
    ## Versioning
    This API uses URL versioning. Current version is v1: `/api/v1/`
  version: 1.0.0
  contact:
    name: Secretly Team
    email: support@secretly.dev
    url: https://secretly.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: API v1
  - url: https://api.secretly.dev/v1
    description: Production server
  - url: https://staging-api.secretly.dev/v1
    description: Staging server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "The request is invalid"
        code:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: object
          description: Additional error details
          example: {"field": "name", "issue": "required"}
    PaginatedResponse:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 100
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20
        total_pages:
          type: integer
          description: Total number of pages
          example: 5
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Authentication required"
            code: 401
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            message: "Insufficient permissions"
            code: 403
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NotFound"
            message: "Resource not found"
            code: 404
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ValidationError"
            message: "Invalid request data"
            code: 400
paths:
...
  /secrets/{id}/share:
    post:
      summary: Share secret
      description: Share a secret with a user or group
      tags: [Secret Sharing]
      parameters:
        - name: id
          in: path
          required: true
          description: Secret ID to share
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareSecretRequest'
      responses:
        '201':
          description: Secret shared successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShareRecord'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /secrets/{id}/shares:
    get:
      summary: List secret shares
      description: List all shares for a specific secret (owner only)
      tags: [Secret Sharing]
      parameters:
        - name: id
          in: path
          required: true
          description: Secret ID
          schema:
            type: integer
          example: 123
      responses:
        '200':
          description: List of shares for the secret
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          shares:
                            type: array
                            items:
                              $ref: '#/components/schemas/ShareRecord'
                          total:
                            type: integer
                            example: 5
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /secrets/{id}/self-share:
    delete:
      summary: Remove self from share
      description: Allow users to remove themselves from a shared secret
      tags: [Secret Sharing]
      parameters:
        - name: id
          in: path
          required: true
          description: Secret ID
          schema:
            type: integer
          example: 123
      responses:
        '204':
          description: Successfully removed from share
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /shares:
    get:
      summary: List user shares
      description: List all shares created by the current user
      tags: [Secret Sharing]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: secret_id
          in: query
          description: Filter by secret ID
          schema:
            type: integer
          example: 123
        - name: recipient_id
          in: query
          description: Filter by recipient ID
          schema:
            type: integer
          example: 456
        - name: permission
          in: query
          description: Filter by permission level
          schema:
            type: string
            enum: [read, write]
          example: read
      responses:
        '200':
          description: List of shares created by the user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          shares:
                            type: array
                            items:
                              $ref: '#/components/schemas/ExtendedShareRecord'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /shares/{id}:
    put:
      summary: Update share permission
      description: Update the permission level for an existing share (owner only)
      tags: [Secret Sharing]
      parameters:
        - name: id
          in: path
          required: true
          description: Share ID
          schema:
            type: integer
          example: 789
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShareRequest'
      responses:
        '200':
          description: Share permission updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShareRecord'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Revoke share
      description: Revoke access to a shared secret (owner only)
      tags: [Secret Sharing]
      parameters:
        - name: id
          in: path
          required: true
          description: Share ID
          schema:
            type: integer
          example: 789
      responses:
        '204':
          description: Share revoked successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /shared-secrets:
    get:
      summary: List shared secrets
      description: List all secrets shared with the current user
      tags: [Secret Sharing]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: permission
          in: query
          description: Filter by permission level
          schema:
            type: string
            enum: [read, write]
          example: read
        - name: owner_id
          in: query
          description: Filter by owner ID
          schema:
            type: integer
          example: 123
      responses:
        '200':
          description: List of secrets shared with the user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          secrets:
                            type: array
                            items:
                              $ref: '#/components/schemas/SharedSecret'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...